+++ packages/vibrant-components/src/lib/BreadCrumbs/index.ts
@@ -1,0 +1,1 @@
+export { BreadCrumbs } from './BreadCrumbs';